move logic into home vc
