+++ GoReceitas/ViewController/Home/Protocols/DefaultCells/DefaultCellsDelegate.swift
@@ -9,4 +9,5 @@ import Foundation
 
 protocol DefaultCellsDelegate: AnyObject {
     func didTapFoodCell(food: FoodResponse)
+    func didFavoriteItem(itemSelected: FoodResponse, favorited: Bool)
 }
