+++ GoReceitas/ViewController/Home/TableViewCells/TryItOut/TryItOutTableViewCell.swift
@@ -6,18 +6,9 @@
 //
 
 import UIKit
-import FirebaseDatabase
-import FirebaseAuth
-
-protocol TryItOutTableViewCellDelegate: AnyObject {
-    func didFavoriteItem(index: IndexPath)
-}
 
 class TryItOutTableViewCell: UITableViewCell {
     private var foodList: [FoodResponse] = [FoodResponse]()
-    private var favoritesArray: [[String : Any]] = [[:]]
-    
-    let auth = Auth.auth()
     
     @IBOutlet weak var pageControl: UIPageControl!
     @IBOutlet weak var collectionView: UICollectionView!
@@ -35,25 +26,6 @@ class TryItOutTableViewCell: UITableViewCell {
         backgroundColor = .viewBackgroundColor
         selectionStyle = .none
         configCollectionView()
-        populateArray()
-    }
-    
-    func populateArray() {
-        if let user = Auth.auth().currentUser {
-            DispatchQueue.global(qos: .userInitiated).async { [weak self] in
-                let database = Database.database().reference()
-                guard let email = user.email else { return }
-                let emailFormatted = email.replacingOccurrences(of: ".", with: "-").replacingOccurrences(of: "@", with: "-")
-                database.child("users/\(emailFormatted)").child("favorites").observeSingleEvent(of: .value, with: { (snapshot) in
-                    // Get the data from the snapshot and assign it to favorites array
-                    if let data = snapshot.value as? [[String : Any]] {
-                        self?.favoritesArray = data
-                    }
-                })
-            }
-        } else {
-            print("There is no currently signed-in user.")
-        }
     }
     
     func configCollectionView() {
@@ -117,34 +89,6 @@ extension TryItOutTableViewCell: DefaultFoodCollectionViewCellDelegate {
         guard let indexPath = collectionView.indexPath(for: cell) else { return }
         let foodSelected = foodList[indexPath.row]
         
-        if let user = Auth.auth().currentUser {
-            DispatchQueue.global(qos: .userInitiated).async {
-                let database = Database.database().reference()
-                guard let email = user.email else { return }
-                let emailFormatted = email.replacingOccurrences(of: ".", with: "-").replacingOccurrences(of: "@", with: "-")
-
-                let favArray: [FoodResponse] = [FoodResponse(id: foodSelected.id, name: foodSelected.name, thumbnail_url: foodSelected.thumbnail_url, cook_time_minutes: foodSelected.cook_time_minutes ?? 0, prep_time_minutes: foodSelected.prep_time_minutes ?? 0, yields: foodSelected.yields ?? "n/a")]
-
-                let mappedArray = favArray.map { ["name": $0.name, "yields": $0.yields ?? "n/a", "image": $0.thumbnail_url, "isFavorited": isActive] }
-
-                // so we can track the array of favorites and update the values correctly
-                self.favoritesArray.append(contentsOf: mappedArray)
-
-                database.child("users/\(emailFormatted)").observeSingleEvent(of: .value, with: { (snapshot) in
-                    if snapshot.hasChild("favorites") {
-
-                        let updates = ["users/\(emailFormatted)/favorites": self.favoritesArray]
-                        database.updateChildValues(updates as [AnyHashable : Any])
-
-                        print("The value already exists in the database.")
-                    } else {
-                        database.child("users/\(emailFormatted)").child("favorites").setValue(mappedArray)
-                        print("The value does not exist in the database.")
-                    }
-                })
-            }
-        } else {
-            print("There is no currently signed-in user.")
-        }
+        delegate?.didFavoriteItem(itemSelected: foodSelected, favorited: isActive)
     }
 }
